requirements.txt
pasgiref==3.3.1
Babel==2.9.0
boto3==1.17.64
botocore==1.20.64
certifi==2020.12.5
chardet==4.0.0
Collectfast==2.2.0
defusedxml==0.6.0
Django==3.0.8
django-admin-menu==1.5
django-allauth==0.42.0
django-ckeditor==6.0.0
django-countries==6.1.3
django-crispy-forms==1.9.1
django-currentuser==0.5.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-phonenumber-field==5.0.0
django-rosetta==0.9.4
django-storages==1.11.1
django-translations==1.1.2
django-versatileimagefield==2.0
jmespath==0.10.0
jsonfield==3.1.0
libsass==0.20.1
oauthlib==3.1.0
phonenumberslite==8.12.16
Pillow==8.1.0
polib==1.1.0
prices==1.1.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-magic==0.4.18
python-magic-bin==0.4.14
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
sentry-sdk==1.1.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.3
uuid==1.30
